Move framework SDK modules to framework core
