+++ api/api/alogging.py
@@ -1,7 +1,7 @@
 # Copyright (C) 2015-2019, Wazuh Inc.
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is a free software; you can redistribute it and/or modify it under the terms of GPLv2
-from wazuh.wlogging import WazuhLogger
+from wazuh.core.wlogging import WazuhLogger
 import logging
 import re
 from aiohttp.abc import AbstractAccessLogger
