+++ framework/wazuh/core/pyDaemonModule.py
@@ -4,11 +4,11 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 
-import errno
 import os
 import sys
 from wazuh import common
 from wazuh.exception import WazuhInternalError
+
 
 def pyDaemon():
     """
@@ -52,6 +52,7 @@ def pyDaemon():
     # Decouple from parent environment
     os.chdir('/')
 
+
 def create_pid(name, pid):
     filename = "{0}/{1}/{2}-{3}.pid".format(common.ossec_path, common.os_pidfile, name, pid)
 
@@ -61,6 +62,7 @@ def create_pid(name, pid):
             os.chmod(filename, 0o640)
         except OSError as e:
             raise WazuhInternalError(3002, str(e))
+
 
 def delete_pid(name, pid):
     filename = "{0}/{1}/{2}-{3}.pid".format(common.ossec_path, common.os_pidfile, name, pid)
