+++ framework/wazuh/core/syscheck.py
@@ -2,7 +2,6 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is free software; you can redistribute it and/or modify it under the terms of GP
 
-import time
 from datetime import datetime
 
 from wazuh.utils import WazuhDBQuery, WazuhDBBackend
