+++ framework/wazuh/core/wdb.py
