+++ framework/wazuh/core/wlogging.py
